+++ src/main/java/com/bookstore/online/domain/book/entity/repository/BooksRepository.java
@@ -4,6 +4,7 @@ import com.bookstore.online.domain.book.entity.BooksEntity;
 import com.bookstore.online.domain.book.entity.resultSet.GetBookOrderCountResultSet;
 import com.bookstore.online.domain.book.entity.resultSet.GetUserOrderPurchasedBookResultSet;
 import java.util.List;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -14,19 +15,25 @@ public interface BooksRepository extends JpaRepository<BooksEntity, Integer> {
 
   List<BooksEntity> findByCategoryNumber(Integer categoryNumber);
   BooksEntity findByBookNumber(Integer bookNumber);
+  List<BooksEntity> findByOrderByBookNameAsc(Pageable pageable);
 
   @Query(value =
       "SELECT * FROM books " +
           "WHERE category_number=:categoryNumber " +
-          "ORDER By :orderSet DESC", nativeQuery = true)
+          "ORDER By :orderSet DESC " +
+          "LIMIT :paging, 10",
+      nativeQuery = true)
   List<BooksEntity> getBookList(@Param("categoryNumber") Integer categoryNumber,
-      @Param("orderSet") String orderSet);
+      @Param("orderSet") String orderSet, @Param("paging") Integer paging);
 
   @Query(value = "SELECT * FROM Books ORDER BY registration_date DESC LIMIT 5", nativeQuery = true)
   List<BooksEntity> getRecentlyBookList();
 
-  @Query(value = "SELECT * FROM books WHERE discount_rate > 0", nativeQuery = true)
-  List<BooksEntity> getBookDisCountList();
+  @Query(value =
+      "SELECT * FROM books WHERE discount_rate > 0 " +
+          "LIMIT :paging, 10"
+      , nativeQuery = true)
+  List<BooksEntity> getBookDisCountList(@Param("paging") Integer paging);
 
   //추루 orders | order_items 기능이 추가 된 이후 주석 해제하여 이용 예정
 //  @Query("SELECT new com.bookstore.online.domain.book.entity.resultSet.GetBookOrderCountResultSet(oi.bookNumber, COUNT(*)) " +
