+++ src/main/java/com/bookstore/online/domain/book/service/ReadBookService.java
@@ -7,6 +7,8 @@ import com.bookstore.online.domain.book.entity.resultSet.GetUserOrderPurchasedBo
 //import com.bookstore.online.domain.orders.OrdersRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -14,27 +16,22 @@ import org.springframework.stereotype.Service;
 public class ReadBookService {
 
   private final BooksRepository booksRepository;
-//  private final OrdersRepository ordersRepository;
-
-  public List<BooksEntity> findCategoryNumber(Integer categoryNumber) {
-    return booksRepository.findByCategoryNumber(categoryNumber);
-  }
 
   public BooksEntity findBookNumber(Integer bookNumber) {
     System.out.println(bookNumber);
     return booksRepository.findByBookNumber(bookNumber);
   }
 
-  public List<BooksEntity> BookList() {
-    return booksRepository.findAll();
+  public List<BooksEntity> BookList(Pageable pageable) {
+    return booksRepository.findByOrderByBookNameAsc(pageable);
   }
 
-  public List<BooksEntity> searchBookList(Integer categoryNumber, String orderSet) {
-    return booksRepository.getBookList(categoryNumber, orderSet);
+  public List<BooksEntity> searchBookList(Integer categoryNumber, String orderSet, Integer paging) {
+    return booksRepository.getBookList(categoryNumber, orderSet, paging);
   }
 
-  public List<BooksEntity> findBookDiscountList() {
-    return booksRepository.getBookDisCountList();
+  public List<BooksEntity> findBookDiscountList(Integer paging) {
+    return booksRepository.getBookDisCountList(paging);
   }
 
   //추후 추가 기능
