+++ src/main/java/com/bookstore/online/domain/category/entity/repository/CategoryRepository.java
@@ -2,11 +2,12 @@ package com.bookstore.online.domain.category.entity.repository;
 
 import com.bookstore.online.domain.category.entity.CategoryEntity;
 import java.util.List;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface CategoryRepository extends JpaRepository<CategoryEntity, Integer> {
-  List<CategoryEntity> findByOrderByCategoryNameAsc();
+  List<CategoryEntity> findByOrderByCategoryNameAsc(Pageable pageable);
   CategoryEntity findByCategoryNumber(Integer categoryNumber);
 }
