+++ src/main/java/com/bookstore/online/domain/category/service/ReadCategoryService.java
@@ -4,6 +4,7 @@ import com.bookstore.online.domain.category.entity.CategoryEntity;
 import com.bookstore.online.domain.category.entity.repository.CategoryRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -11,8 +12,8 @@ import org.springframework.stereotype.Service;
 public class ReadCategoryService {
   private final CategoryRepository categoryRepository;
 
-  public List<CategoryEntity> readCategoryAllList(){
-    return categoryRepository.findByOrderByCategoryNameAsc();
+  public List<CategoryEntity> readCategoryAllList( Pageable pageable) {
+    return categoryRepository.findByOrderByCategoryNameAsc(pageable);
   }
 
   public CategoryEntity findCategoryNumber(Integer categoryNumber){
