+++ src/main/java/com/bookstore/online/domain/user/controller/ExceptionControllerAdvice.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
-public class ExceptionControllerAdvice {
+public class  ExceptionControllerAdvice {
 
   @ExceptionHandler(SignupValidException.class)
   public ResponseEntity<?> validException(SignupValidException e) {
